[Session094] refactor(#2): extract push-notification subsystem into usePushNotifications hook
First behavior-preserving slice of the AppContent god-component refactor.
Moves the Web Push subscription state, per-type preference toggles, and the
Edge-Function send helpers (subscribe/unsubscribe/toggle/send/test) out of
AppContent into src/hooks/usePushNotifications.js. The hook receives the
cross-cutting deps (supabase, user, leagueId, showToast) and returns the same
values/functions AppContent previously exposed, so SettingsView and all push
consumers are unchanged. urlBase64ToUint8Array and the VAPID import move with
it. AppContent drops ~170 lines (1798 -> 1627). Verified in preview: app
mounts, Settings renders all notification toggles, no console errors.
